pantsu/pomf: Add "disable issues" task for 1.1.0
Gogs' issue tracker is centralized. I see this as unethical, because it
invites users to vendor lock-in themselves to Pantsu.cat's issue
tracker.

The proposal to address this issue is to use a mailing list, which is
more decentralized/distributed and every client can do their computing
on their own computers.