sysadmin: Add "distribute services to IPv4" task
A single IP-address is the single point of failure, e.g. in case of a
nullroute.

HTTPS also has to use SNI to detect which site the user wants to go to.
This is unsupported in legacy browsers, although the site's TLS
configuration does not support common legacy browsers anyway.