pantsu/pomf: Move upload_file() task to 1.1.0
I don't want to rush with this release. Keeping this task in 1.1.1 is
bad because it would be an intentional "bug" left in the 1.1.0 release
otherwise.

pantsu/pomf@8425a1bd35 partially addresses this concern. However it was
decided on IRC that this is not a pretty way to fix it and the code will
be refactored first a little bit.