pantsu/pomf: Add "Replace JSHint with ESLint" for 1.1.0
JSHint is actively deprecating style checks and dedicating them to JSCS.

I am personally not in favor of supporting two different tools, even
though I am all for the UNIX philosophy. Still, ESLint is faster and has
better ES6 support (and supports JSX).

The JSHint task has been merged to pantsu/pomf master branch for release
1.1.0, however this seems like a bad idea now. The suggestion is to
replace it with ESLint before it makes an appearance in the release, so
that we don't have to support JSHint after 1.1.0.