pantsu/pomf: Add "fix pomf.js violations" for 1.1.0
I believe we tried to fix the three "JSHint" violations earlier, but we
actually hit another bug on the road with an addition symbol
(pantsu/pomf@7b09c76c8d). This means our attempts to fix the violations
may have failed for unrelated reasons.

I think another attempt should be given at this before release to assure
the best quality for this release.