pantsu/pomf: Add "semantic <button>" task for 1.1.0
The index page has an upload button, which is commonly large and green
and says "Select or drop file(s)". It uses an <a> element and works like
a <button> element.

This has been my pet-peeve for a long time. Anchors (<a>) are not
buttons, they are hyperlinks which link to other content. The idea to
use <a> for buttons comes from years ago when you couldn't style
<button> elements. Today you can, so there's no reason to break the
semantics. It is bad for accessibility.

I believe we tested a simple element change from <a> to <button> to be
working, but I want to make sure and test this again with ewhal@ to make
sure before merging it to master.

Reference: http://blogs.ancestry.com/techroots/buttons-vs-anchors/