pantsu/pomf: Add "jQuery to vanilla JS" for 3.0.0
There's some inconsistencies in pomf.js. Some things are done in jQuery
when they could be done in pure JavaScript easily, and vice versa.

For portability and perhaps even performance reasons, it's better to aim
for pure JavaScript. This goal specifically aims 3.0.0 release to be
written in pure JavaScript if possible.