pantsu/pomf: Move "/pages/en/" task to 3.0.0
It's too early to make the change. We haven't announced URLs to change
yet as a deprecation warning. The other thing blocking this is
js/pomf.js, which assumes there is always upload.php in same folder
root. The result is duplicated code, which is difficult to maintain
without an API.