pantsu/pomf: Add fputcsv() task to 2.0.0
This concerns wubthecaptain/pomf@363f107 .

The code for CSV responses in Response class is not very ideal. It is
difficult to read the escaped double quotes, while we could use arrays
and fputcsv() to improve code readability.

If this does not make it into 2.0.0 release, then it should be fixed in
2.0.1.