pantsu/pomf: Add `bannersca.json` task for 2.0.0
grunt-swig has a config option `banners:`. This is configured to use
`pages/banners.json`. There is a bug where Catalan translated banners
are not displayed on Catalan pages. Including both banner files in
`Gruntfile.js` results both Catalan banners and English banners being
displayed on English pages.

In other words, this is not a proper way to do i18n. It's buggy and it
results in code duplication.