pantsu/pomf: Remove i18n subfolders task from 3.0.0
The subfolders task was to fix a problem where all translated pages
would be located in a messy /pages/ folder, such as /pages/index.swig
and /pages/indexca.swig. This old, poorly thought i18n support has been
removed in pantsu/pomf@566611c.

With future gettext implementations, there will be no more duplicate
Swig templates so there are no pages to sort to subfolders. Instead,
there will be an i18n folder (or an extension) with a bunch of
.pot/.po/.mo files.