pantsu/pomf: Remove "rewrite CSV to use array()"
wubthecaptain/pomf@73fbb97 acknowledges this to be a future improvement,
if it ever happens to be necessary. New discussion has been started on
IRC if we even want to support the CSV response at all. Rewriting CSV is
not a required goal for 2.0.0 and can be addressed in a patch or minor
version increment.