pantsu/pomf: Add code review tasks for 2.0.0
Add `wubthecaptain/pomf` branches `add-git-mailmap` and `cleanup-css`
for code review.

Pomf development is stalling because of the lead developer's (ewhal@)
absence. Because nobody on IRC is ready to do actual code review, they
do it with vague OKs only or sometimes incorrectly give OKs for branches
with known bugs, code reviews are now a TODO task that need attention
badly.