pantsu/pomf@2.0.0: Add bugfix task for empty response
Some things were shuffled in the refactor-response-api branch, which is
pretty big already. There's a bug with Gyazo responses being empty that
remains unfixed, thanks to nobody volunteering to test and suggest
patches (I'm looking at you, ewhal@).

Because no attention has been given to this task for 3 weeks or more,
it's now a TODO task that needs attention and which anyone can pick up.