sysadmin: Remove gogs issue tracker task
git.pantsu.cat used to run on an old version of gogs earlier when no
feature existed to disable issue tracking. We wanted to disable issue
tracking because it turned out to be a walled garden and generally a bad
idea for user's freedom, versus a mailing list.

Right now git.pantsu.cat is running gogs 0.8.15.1231. Sometime in the
0.8.x branch this feature was introduced, and I disabled issue trackers
about a week or two ago.

Thus, this task is complete and can be removed.