sysadmin: Add cgit task to replace gogs
gogs is full of bugs that are a nuisance to me personally, and the
social features of gogs are not really a good idea; they can easily turn
into a walled garden.

cgit on other hand is really fast and used by large projects like the
Kernel Linux. It is no-frills added, does exactly the one thing it
should do and not ten other things with it.