pantsu/pomf@3.0.0: Add application/vnd.api+json task
The JSON API specification mandates the use of application/vnd.api+json
media type for JSON APIs. IANA also recognizes this media type. Pomf is
currently using application/json, which is technically correct but not
compliant with the standard.

This task is for 3.0.0 because of potential backwards compatibility
concerns. Most existing tools like ShareX don't seem to care, but other
unknown clients using Accept HTTP headers may be impacted. Pomf also
doesn't use API versioning, so this is what we have to do for now.