pantsu/pomf@2.1.0: Ask if our build system supports ES6
Luminarys reported Grunt throwing errors with ES6 code, specifically
with let variables. I don't have more details, but this needs further
investigation.

If we are going to need a new build system, this task will be pushed
back to pantsu/pomf@3.0.0. It may be a good idea anyway due to potential
backwards compatibility concerns with browser support for ES6.