pantsu/pomf@2.1.0: Add #upload-btn replacement task
CSS IDs and classes should be used for styling only, not for JavaScript
hooks. Unfortunately, we are using `#upload-btn` as a JavaScript hook at
the moment. This will be trivial to correct to `.js-upload-btn` to be
explicit there are no styles associated with the hook.

As far as I am aware, there are no backwards compatibility issues with
known third-party Pomf stylesheets.