pantsu/pomf@2.1.0: Add a task to lint PHP again
Looks like I didn't run PHP Mess Detector (phpmd) on the source code
last time at @2.0.0 release. It may also be that linting to PSR-2 was
incomplete or never merged.

For example, PHPMD gives the following warnings:

    The method upload_file uses an else expression. Else is never
    necessary and you can simplify the code to work without else.

May be merged to 2.0.x maint branch if we can be sure it won't break
things.