pantsu/pomf@2.1.0: Add label/title task for index.swig
Somewhere around commit 6f51b356a4c08f8ec3b452e92e301062c828dbc5, the
intention was to improve accessibility of the uploader. However,
implementing a visually "non-disruptive" <label> element for the
uploader proved to be difficult because we use a custom <button> element
instead of the browser's default <input> for interface.

This accessibility bug was then forgotten and never revisited. But have
no fear, I got you this time and remembered to check.

nojs.swig is also affected, but I estimate that to be easier to fix.