pantsu/pomf: Postpone UTF-8 and try{} to 2.2.0
It feels to me like there's too much to do for 2.1.0 again. Let's do a
smaller release, even if something like UTF-8 support should be easy to
implement.

This way, I can focus on fixing front end bugs, cleanup and linting the
base in 2.1.0 before working on back end PHP bugs.

Hopefully this way the amount of bugs that go into a release will be
also smaller, reducing the amount of maint branch releases and
backporting.

All features are misfeatures.