pantsu/pomf@2.2.0: Remove UTF-8 file name task
Although we could add UTF-8 support for the MySQL schema, there seems to
be no real benefit to this after version 3.0.0. Existing databases would
also have trouble doing the migration.

The (yet to be listed) plan for Pomf 3.0.0 is to remove all the unused
and useless MySQL schema values, one of which is originalname. That
happens to be the only value that would sanely benefit from UTF-8
support.

I suggest the MySQL or MariaDB and PHP authors take steps to make UTF-8
a default, all things better concerned.

Discussed with Austin Gillmann <alucard@cuntflaps.me>.